refactor: create module for wait subcommand
